Combine PComparer and PScorer into just PScorer
PScorer is now more abstract, to allow
for multiple types of scoring:
Feature-based scoring, which is the classical technique, composed
of generating pairwise features and
then classifying them.
But there is also rule-based scoring, which might not
rely on these features.

This also changes the nomenclature from cluster to
partiion, because that is more in line with what we are doing.

It also renames pipeline to Deduper