Return conatiner Blocking and Scoring instead of concrete data
The blocker and scorer might return results that are larger than
memory, so we want to return an abstraction
to iterate over the results.
In the simple case you could just call eg
Blocking.to_arrow() or maybe in the future Blocking.to_pandas(),
but this also allows us to do something like
iter_arrow() and deal with things out of core.