Remove Dataset indirection
Now we operate on raw ibis Tables. There is the implicit
idea that there must always be a unique column called "record_id".
They _can_ contain a column called "label_true" with the true label of
the record.

Users have to follow this convention, it
will make sharing and talking about code easier, not
to mention the implementation in this
lib