Remove DatasetPair, replace with Task
I think we can get away with a lighter solution, of just
using the raw left and right ibis Tables.
We can hold the Dedupe and Link specific stuff,
such as n_possible_pairs and redundant_comparisons()
in DedupeTask and LinkTask classes.
This allows the library functions
such as block() to be lighter weight
and accept raw tables.