feat: remove FSComparer and prior just use Weights
Remove the prior because
- it was arbitrary. users are just going to need to pick
a combined odds to use as a threshold, and the prior just acted as
a scaling of this.
reduce the number of things users must think about/understand.
- IT was more code.

Also remove the FS Comparer class.
Now the Weights are the ones that contain
the `score` method, which I think makes more sense.
They take an already-compared table, so it is more
composable that way as well. Ther responsibilities are
clearer