remove `how` kwarg from label_pairs()
Now this always labels by name.

I only included the how so that you could label by index, because I thought that would be more performant. But that
was a mistake, because even if you labeled by index, then you would
probably then need to follow up with another
mapping to a ComparisonWeight or seomthign.
This would then be two chained CASE
statements. REally these should get optimized into one.

So, the user should be the one to do this inlining themselves. Since the public API of these dataclasses is so simple, this shouldn't
be that burdensome, and it will
lead to better performance.