fix: remove .cache() from sample_table()
This was materializing the whole table!
During weight training, when using
train_us_using_sampling(), we were doing a cross product and sampling from that.
But this was leading to the entire corss product being materialized!
Not good. I'm guessing this will fix (or at least
be in the right direction)
https://github.com/NickCrews/mismo/issues/21