WIP: block: de-class blocking API
This isn't totally done yet, but the changes
are stacking up and I want to get them committed.

I don't like the Class-heavy API before.
It forced you to create BlockingRule or BlockingRules instances.
I would prefer to use a more protocol based approach: you call the
top-level block() with two tables, and some objects that are somehow able to block,
and it figures it out.

Each blocker can either be an atomic.simple rule like an
str (representing a column name), and ibis expression, a literal True,
a lambda.
Or it could be a more complex user-defined object or class. It just has to be callable with the signature (Table, Table, **kwargs) -> <one of those attomics>