deps: update deps
Also reformat with new version of ruff.
This updates dateutils to >2.9.0, which
gets us around [a deprecationwarning](https://github.com/dateutil/dateutil/releases/tag/2.9.0) in python 3.12
that was causing failures in CI

Also pin duckdb away from 0.10.1 and 0.10.2 because I am hitting
https://github.com/duckdb/duckdb/issues/11542

Also add in a workaround for ibis's .bind()
no longer working as before.
This is an internal API of theres, so no blame there.