feat: overhaul sample_all_pairs()
fixes https://github.com/NickCrews/mismo/issues/35

This implementation should be unbiased and fairly performant.
The biggest drawback of it is that it doesn't allow for a random seed. I had to rip it out from here, and then
I went up the callstack and removed it from the callers as well
so we didn't pretend things were
reproducible when they weren't.

I figured that the absolute performance blocker
mentioned in the above issue was more important
to fix than the reproducible seed.

Also make this method public, and add a decent docstring.