privatize address featurization, make actually performant
The API is much worse now, but that is a price we have to pay.
I want this to be private so I'm more free to play around with it as needed.

by doing all these operations in a much of sequential .mutate()s,
it compiles to a bunch of chained CTEs in the SQL.
If we only did one big .mutate(), some of the deep subexpressions would
get copy-pasted many many times (eg Array.unique() results in the argument
being copy-pasted 4 times: # https://github.com/ibis-project/ibis/blob/75ecf038d917b968dcc493c69429abe1e8549dd2/ibis/backends/sql/compilers/duckdb.py#L145-L155).
If duckdb/the backend were clever with common subexpression elimination,
this might not be a problem. But, duckdb appears to evaluate a regex
for every time it appears in the SQL. See https://github.com/duckdb/duckdb/discussions/14649.
So, if we did one .mutate(), we would end up with like literally 100 regex
evaluations in the SQL, which is 100x slower than evaluating the regex once.